third_party/WebKit/LayoutTests/fast/events/popup-blocking-timers6-expected.txt

Click Here
PASS successfullyParsed is true

TEST COMPLETE
Test calling window.open() with a 1001 ms delay. A popup should not be allowed.
PASS newWindow is null
